### 2023-11-16 #6064 @editorialbot
[REVIEW]: scida: scalable analysis for scientific big data
**Submitting author:** @cbyrohl (<a href="http://orcid.org/0000-0002-0885-8090">Chris Byrohl</a>)
**Repository:** https://github.com/cbyrohl/scida
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.3.1
