### 2023-11-20 7b180af
add tutorial selection to install page
