+++ pyproject.toml
@@ -56,6 +56,7 @@ jupyter = "^1.0.0"
 pytest-mock = "^3.11.1"
 notebook = "^6.4.12"
 mkdocs-glightbox = "^0.3.4"
+markdown-include = "^0.8.1"
 
 [tool.coverage.paths]
 source = ["src", "*/site-packages"]
