### 2023-12-04 cbyrohl/scida#98 @EGaraldi
Filename in Tutorial->simulations
The code snippet and sample outputs in the [simulation tutorial](https://cbyrohl.github.io/scida/tutorial/simulations/) do not match the one obtained following the [link](https://heibox.uni-heidelberg.de/f/dc65a8c75220477eb62d/) provided in the same page. The former is "TNG50-4_snapshot" while the latter is "snapdir_030". Similarly, the content of the file is slightly different than showed in the tutorial page. If possible, homogenize them for clarity.
