### 2023-12-04 cbyrohl/scida#99 @EGaraldi
API doc for scida.load()
I couldn't fine the API entry for load(). I also think it would be useful to mention in the tutorial what is accepted by load (i.e. a filename, a directory, ...)
