### 2023-12-04 cbyrohl/scida#100 @EGaraldi
">>>" in code snippets
In the tutorial, some of the code snippets provided have the code prepended by ">>> " making it hard to copy-paste the code into the terminal.
