### 2023-12-04 cbyrohl/scida#101 @EGaraldi
Error when loading GAIA dataset in tutorial
When loading GAIA dataset I get the following errors
INFO:scida.interfaces.mixins.units:Cannot parse unit string 'electron'.s**-1' from metadata description. Skipping.
INFO:scida.interfaces.mixins.units:Cannot parse unit string 'mag' from metadata description. Skipping.
INFO:scida.interfaces.mixins.units:Cannot parse unit string 'electron'.s**-1' from metadata description. Skipping.
INFO:scida.interfaces.mixins.units:Cannot parse unit string 'mag' from metadata description. Skipping.
INFO:scida.interfaces.mixins.units:Cannot parse unit string 'electron'.s**-1' from metadata description. Skipping.
INFO:scida.interfaces.mixins.units:Cannot parse unit string 'mag' from metadata description. Skipping.
