### 2023-12-04 cbyrohl/scida#102 @EGaraldi
2D histogram from GAIA looks different when running the code
The 2D histogram from GAIA (in Tutorial->observations) is different than what you get by running the code above it (maybe the plot shown is for the full GAIA dataset and not a subset of it? If so, place a warning)
