### 2023-12-04 cbyrohl/scida#103 @EGaraldi
Typo in FITS files documentation
In the first line, I believe it should read "Support in scida **is** work-in-progress"
