### 2023-12-04 cbyrohl/scida#104 @EGaraldi
Errors in code in docs unit cookbook entry
The code in the [units cookbook](https://cbyrohl.github.io/scida/notebooks/cookbook/units/) does not execute and displays errors.
