### 2023-12-04 cbyrohl/scida#105 @EGaraldi
Missing variable names in API docs
In the API docs, the boxes describing the function parameters are not showing the variable name they are referring to, only its type.
