### 2023-12-04 cbyrohl/scida#106 @EGaraldi
JOSS paper: missing comparison with tangos
I suggest to include a short comparison with [tangos](https://pynbody.github.io/tangos/) concerning the load/linking of variations of the same simulations. Are scida and tangos alternative to each other (in this particular task)? Do they work together? Neither?
