### 2023-12-04 #6064 @EGaraldi
I opened a number of issues in the repository, namely:

- [Filename in Tutorial->simulations](https://github.com/cbyrohl/scida/issues/98)
- [API doc for scida.load()](https://github.com/cbyrohl/scida/issues/99)
- [">>>" in code snippets](https://github.com/cbyrohl/scida/issues/100)
- [Error when loading GAIA dataset in tutorial](https://github.com/cbyrohl/scida/issues/101)
- [2D histogram from GAIA looks different when running the code](https://github.com/cbyrohl/scida/issues/102)
- [Typo in FITS files documentation](https://github.com/cbyrohl/scida/issues/103)
- [Errors in code in docs unit cookbook entry](https://github.com/cbyrohl/scida/issues/104)
- [Missing variable names in API docs](https://github.com/cbyrohl/scida/issues/105)
- [JOSS paper: missing comparison with tangos](https://github.com/cbyrohl/scida/issues/106)

Except for these issue, I find the code very useful and potentially quite impactful.
