### 2023-12-08 0490aee
update doc tool versions + fix #105
