### 2023-12-08 cdc57b9
fix #100 by limiting copy-paste selection
