### 2023-12-08 cbyrohl/scida#100 @cbyrohl
I would still favor pycon mode (implying the use of ">>>") to reduce number of code cells (i.e. splitting input and output), as done in e.g. the [scipy user guide](https://docs.scipy.org/doc/scipy/tutorial/integrate.html)).  We can manipulate the css to never select the ">>>" when manually copy & pasting using the mouse selection, see [here](https://mkdocstrings.github.io/recipes/#prevent-selection-of-in-python-code-blocks), to make the copy-paste more hassle free. Unfortunately, this will not be possible for the copy&paste button (see [here](https://github.com/squidfunk/mkdocs-material/issues/3647)), so I deactivated the button for pycon cells in cdc57b95856877f330c355c92d519c722668a699
