### 2023-12-08 cbyrohl/scida#101 @cbyrohl
This should just be info output (rather than an error; due to units that cannot be parsed by pint), in parts due to limited [logarithmic unit support](https://pint.readthedocs.io/en/stable/user/log_units.html) and unknown unit 'electron' (?), so we will have to think about how to handle this for now. I would keep the verbosity to make the user aware of limited units for these fields.
