### 2023-12-08 cbyrohl/scida#104 @cbyrohl
Page is outdated and will be deleted; instead linked new [unit](https://github.com/cbyrohl/scida/blob/main/docs/units.md) in table of contents.
