### 2023-12-08 cbyrohl/scida#98 @kyleaoman
Likewise the projection image from the example file differs a bit from the one in the documentation.
