### 2023-12-08 221c419
fix #102 by adding info box
