### 2023-12-10 8f9a557
update docs (#107)
* update doc tool versions + fix #105

* fix deprecation warn

* add copy-paste button

* remove unit cookbook

* fix #100 by limiting copy-paste selection

* fix #103

* fix #102 by adding info box

* fix #108
