+++ pyproject.toml
@@ -30,10 +30,10 @@ coverage = "^7.0.1"
 pytest-xdist = {extras = ["psutil"], version = "^3.1.0"}
 pytest-profiling = "^1.7.0"
 mypy = ">=0.930"
-mkdocs = "^1.3.0"
-mkdocs-material = "^8.0.0"
-mkdocstrings = {version = "^0.19.0", extras = ["python-legacy"]}
-mkdocs-jupyter = "^0.22.0"
+mkdocs = "^1.5.3"
+mkdocs-material = "^9.5.1"
+mkdocstrings = {extras = ["python"], version = "^0.24.0"}
+mkdocs-jupyter = "^0.24.6"
 ipykernel = "^6.19.4"
 matplotlib = "^3.6.2"
 ipywidgets = "^8.0.4"
