### 2023-12-10 4c29426
update hist2d example to reflect image
