### 2023-12-11 cbyrohl/scida#101 @EGaraldi
I see. I think at the very least there should be a comment in the tutorial explaining this behavior is understood and expected for the dataset. Maybe the INFO can be made more explicit by saying something like "This unit will not be available" rather than just "Skipping"
