### 2023-12-12 #6064 @mbobra
> @mbobra could you just clarify the action to take now? Is the idea to iterate with the authors on any open issues that I feel would preclude accepting the submission until I'm satisfied?

Yes. You don't need to take the initiative here, though -- @cbyrohl, when you've finished addressing @kyleaoman's open issues, please let us know here. Thank you both!
