### 2023-12-12 cbyrohl/scida#104 @cbyrohl
Thanks a lot, missed that. Fixed with PR #119
