### 2023-12-12 98a0820
series/utilities docstrings
