+++ src/scida/series.py
@@ -1,3 +1,8 @@
+"""
+This module contains the base class for DataSeries, which is a container for collections of dataset instances.
+"""
+
+
 import inspect
 import json
 import os
@@ -18,8 +23,27 @@ from scida.registries import dataseries_type_registry
 
 
 def delay_init(cls):
+    """
+    Decorate class to delay initialization until an attribute is requested.
+    Parameters
+    ----------
+    cls:
+        class to decorate
+
+    Returns
+    -------
+    Delay
+
+    """
+
     class Delay(cls):
+        """
+        Delayed initialization of a class. The class is replaced by the actual class
+        when an attribute is requested.
+        """
+
         def __init__(self, *args, **kwargs):
+            """Store arguments for later initialization."""
             self._args = args
             self._kwargs = kwargs
 
@@ -52,13 +76,26 @@ def delay_init(cls):
             return getattr(self, name)
 
         def __repr__(self):
+            """Return a string representation of the lazy class."""
             return "<Lazy %s>" % cls.__name__
 
     return Delay
 
 
 class DatasetSeries(object):
-    """A container for collection of interface instances"""
+    """A container for collections of dataset instances
+
+    Attributes
+    ----------
+    datasets: list
+        list of dataset instances
+    paths: list
+        list of paths to data
+    names: list
+        list of names for datasets
+    hash: str
+        hash of the object, constructed from dataset paths.
+    """
 
     def __init__(
         self,
@@ -71,6 +108,27 @@ class DatasetSeries(object):
         names=None,
         **interface_kwargs
     ):
+        """
+
+        Parameters
+        ----------
+        paths: list
+            list of paths to data
+        interface_args:
+            arguments to pass to interface class
+        datasetclass:
+            class to use for dataset instances
+        overwrite_cache:
+            whether to overwrite existing cache
+        lazy:
+            whether to initialize datasets lazily
+        async_caching:
+            whether to cache datasets asynchronously. TODO: not used; delete?
+        names:
+            names for datasets
+        interface_kwargs:
+            keyword arguments to pass to interface class
+        """
         self.paths = paths
         self.names = names
         self.hash = hash_path("".join([str(p) for p in paths]))
@@ -109,16 +167,51 @@ class DatasetSeries(object):
             pass
 
     def __init_subclass__(cls, *args, **kwargs):
+        """
+        Register datasetseries subclass in registry.
+        Parameters
+        ----------
+        args:
+            (unused)
+        kwargs:
+            (unused)
+        Returns
+        -------
+
+        """
         super().__init_subclass__(*args, **kwargs)
         dataseries_type_registry[cls.__name__] = cls
 
     def __len__(self):
+        """Return number of datasets in series.
+
+        Returns
+        -------
+        int
+        """
         return len(self.datasets)
 
     def __getitem__(self, key):
+        """
+        Return dataset by index.
+        Parameters
+        ----------
+        key
+
+        Returns
+        -------
+
+        """
         return self.datasets[key]
 
     def info(self):
+        """
+        Print information about this datasetseries.
+
+        Returns
+        -------
+        None
+        """
         rep = ""
         rep += "class: " + sprint(self.__class__.__name__)
         props = self._repr_dict()
@@ -154,7 +247,13 @@ class DatasetSeries(object):
         print(rep)
 
     @property
-    def data(self):
+    def data(self) -> None:
+        """
+        Dummy property to make user aware this is not a Dataset instance.
+        Returns
+        -------
+
+        """
         raise AttributeError(
             "Series do not have 'data' attribute. Load a dataset from series.get_dataset()."
         )
@@ -174,7 +273,7 @@ class DatasetSeries(object):
 
     def __repr__(self) -> str:
         """
-        Return a string representation of the object.
[…]
