### 2023-12-13 356144a
Update CONTRIBUTING.md; typo
