+++ CONTRIBUTING.md
@@ -1,3 +1,3 @@
 We welcome contributions to scida in the form of bug reports, feature requests, design proposals and more.
 
-For information on how to contribute, please see the https://cbyrohl.github.io/scida/develop.
+For information on how to contribute, please see the https://cbyrohl.github.io/scida/developer/.
