### 2024-01-02 ef89703
fix #115 by adjusting example
