### 2024-01-02 1679b21
fix #115  (#122)
* fix #115 by adjusting example + add pint-dask user warning to faq
