### 2024-01-02 cbyrohl/scida#98 @cbyrohl
Thanks @EGaraldi, non standard name dropped in favor of folder name "snapdir_030" in #126 . Thanks @kyleaoman, field reference and projection image adjusted in #114
