### 2024-01-03 68c8dbb
Update FLAMINGO test for index units
