### 2024-01-03 6936161
improve docs on failing unit inference
