+++ tests/customs/test_flamingo.py
@@ -45,7 +45,7 @@ def check_flamingosnap(obj: SwiftSnapshot, obj_wu: SwiftSnapshot):
         assert pytest.approx(uq) == uq_ref
         assert v.units == cgs_units[k]
 
-    for ptype, d in obj.data.items():
+    for d in obj.data.values():
         for k in d.keys():
             if not k.endswith("IDs"):
                 continue
