### 2024-01-03 4709b9e
Unit inference reporting (#130)
* Update FLAMINGO test for index units

* Improve unit reporting

* improve docs on failing unit inference

* add missing tng sim doc page
