+++ src/scida/misc.py
@@ -1,8 +1,3 @@
-"""
-Miscellaneous helper functions.
-"""
-
-
 import logging
 import os
 import pathlib
@@ -22,20 +17,6 @@ log = logging.getLogger(__name__)
 def get_container_from_path(
     element: str, container: FieldContainer = None, create_missing: bool = False
 ) -> FieldContainer:
-    """
-    Get a container from a path.
-    Parameters
-    ----------
-    element: str
-    container: FieldContainer
-    create_missing: bool
-
-    Returns
-    -------
-    FieldContainer:
-        container specified by path
-
-    """
     keys = element.split("/")
     rv = container
     for key in keys:
@@ -49,21 +30,8 @@ def get_container_from_path(
     return rv
 
 
-def return_hdf5cachepath(path, fileprefix: Optional[str] = None) -> str:
-    """
-    Returns the path to the cache file for a given path.
-    Parameters
-    ----------
-    path: str
-        path to the dataset
-    fileprefix: Optional[str]
-        Can be used to specify the fileprefix used for the dataset.
-
-    Returns
-    -------
-    str
-
-    """
+def return_hdf5cachepath(path_original, fileprefix=None) -> str:
+    path = path_original
     if fileprefix is not None:
         path = os.path.join(path, fileprefix)
     hsh = hash_path(path)
@@ -72,19 +40,7 @@ def return_hdf5cachepath(path, fileprefix: Optional[str] = None) -> str:
 
 
 def path_hdf5cachefile_exists(path, **kwargs) -> bool:
-    """
-    Checks whether a cache file exists for given path.
-    Parameters
-    ----------
-    path:
-        path to the dataset
-    kwargs:
-        passed to return_hdf5cachepath
-    Returns
-    -------
-    bool
-
-    """
+    """Checks whether a cache file exists for given path."""
     fp = return_hdf5cachepath(path, **kwargs)
     if os.path.isfile(fp):
         return True
@@ -92,17 +48,7 @@ def path_hdf5cachefile_exists(path, **kwargs) -> bool:
 
 
 def return_cachefile_path(fname: str) -> Optional[str]:
-    """
-    Return the path to the cache file, return None if path cannot be generated.
-    Parameters
-    ----------
-    fname: str
-        filename of cache file
-
-    Returns
-    -------
-
-    """
+    """If path cannot be generated, return None"""
     config = get_config()
     if "cache_path" not in config:
         return None
@@ -118,19 +64,8 @@ def return_cachefile_path(fname: str) -> Optional[str]:
     return fp
 
 
-def map_interface_args(paths: list, *args, **kwargs):
-    """
-    Map arguments for interface if they are not lists.
-    Parameters
-    ----------
-    paths
-    args
-    kwargs
-
-    Returns
-    -------
-
-    """
+def map_interface_args(paths, *args, **kwargs):
+    """Map arguments for interface if they are not lists"""
     n = len(paths)
     for i, path in enumerate(paths):
         targs = []
@@ -149,18 +84,6 @@ def map_interface_args(paths: list, *args, **kwargs):
 
 
 def str_is_float(element: str) -> bool:
-    """
-    Check whether a string can be converted to a float.
-    Parameters
-    ----------
-    element: str
-        string to check
-
-    Returns
-    -------
-    bool
-
-    """
     try:
         float(element)
         return True
@@ -171,31 +94,6 @@ def str_is_float(element: str) -> bool:
 def rectangular_cutout_mask(
     center, width, coords, pbc=True, boxsize=None, backend="dask", chunksize="auto"
 ):
-    """
[…]
