### 2024-01-03 da3c322
Revert "Revert "docstrings for io and misc""
This reverts commit 2c87a2c63750489036c4ca78bb57de7020b03869.
