### 2024-01-03 f82742e
docstrings for config.py
