### 2024-01-03 5123de9
docstrings for interface.py
