### 2024-01-03 f8a26a3
improve docstrings
