+++ docs/api/moduleindex.md
@@ -1,0 +1,5 @@
+# Advanced API
+
+For now, we list the remaining package documentation here.
+
+:::scida
