+++ pyproject.toml
@@ -31,7 +31,7 @@ pytest-xdist = {extras = ["psutil"], version = "^3.1.0"}
 pytest-profiling = "^1.7.0"
 mypy = ">=0.930"
 mkdocs = "^1.5.3"
-mkdocs-material = "^9.5.1"
+mkdocs-material = "^9.5.3"
 mkdocstrings = {extras = ["python"], version = "^0.24.0"}
 mkdocs-jupyter = "^0.24.6"
 ipykernel = "^6.19.4"
