### 2024-01-03 1759430
minor fix for unit mismatch check
