### 2024-01-18 da690fb
add fieldcontainer copy func (#133)
