### 2024-01-29 5ff5f1f
misc docstrings
