+++ src/scida/convenience.py
@@ -86,6 +86,7 @@ def download_and_extract(
 def get_testdata(name: str) -> str:
     """
     Get path to test data identifier.
+
     Parameters
     ----------
     name: str
@@ -95,7 +96,7 @@ def get_testdata(name: str) -> str:
     str
     """
     config = get_config()
-    tdpath = config.get("testdata_path", None)
+    tdpath: Optional[str] = config.get("testdata_path", None)
     if tdpath is None:
         raise ValueError("Test data directory not specified in configuration")
     if not os.path.isdir(tdpath):
@@ -109,6 +110,7 @@ def get_testdata(name: str) -> str:
 def find_path(path, overwrite=False) -> str:
     """
     Find path to dataset.
+
     Parameters
     ----------
     path: str
@@ -313,6 +315,7 @@ def load(
 def get_dataset_by_name(name: str) -> Optional[str]:
     """
     Get dataset name from alias or name found in the configuration files.
+
     Parameters
     ----------
     name: str
@@ -341,6 +344,19 @@ def get_dataset_by_name(name: str) -> Optional[str]:
 
 
 def get_datasets_by_props(**kwargs):
+    """
+    Get dataset names by properties.
+
+    Parameters
+    ----------
+    kwargs: dict
+        Properties to match.
+
+    Returns
+    -------
+    list[str]:
+        List of dataset names.
+    """
     dnames = []
     c = get_config()
     if "datasets" not in c:
@@ -362,6 +378,22 @@ def get_datasets_by_props(**kwargs):
 
 
 def get_dataset_candidates(name=None, props=None):
+    """
+    Get dataset candidates by name or properties.
+
+    Parameters
+    ----------
+    name: Optional[str]
+        Name or alias of dataset.
+    props: Optional[dict]
+        Properties to match.
+
+    Returns
+    -------
+    list[str]:
+        List of candidate dataset names.
+
+    """
     if name is not None:
         dnames = [get_dataset_by_name(name)]
         return dnames
@@ -372,6 +404,21 @@ def get_dataset_candidates(name=None, props=None):
 
 
 def get_dataset(name=None, props=None):
+    """
+    Get dataset by name or properties.
+    Parameters
+    ----------
+    name: Optional[str]
+        Name or alias of dataset.
+    props: Optional[dict]
+        Properties to match.
+
+    Returns
+    -------
+    str:
+        Dataset name.
+
+    """
     dnames = get_dataset_candidates(name=name, props=props)
     if len(dnames) > 1:
         raise ValueError("Too many dataset candidates.")
