### 2024-01-29 4840bb9
dataset/series docstrings
