### 2024-01-29 eac32e7
clarify path argument for load().
