+++ src/scida/customs/arepo/MTNG/dataset.py
@@ -1,3 +1,7 @@
+"""
+Support for MTNG-Arepo datasets, see https://www.mtng-project.org/
+"""
+
 import os
 from typing import Union
 
@@ -15,6 +19,20 @@ class MTNGArepoSnapshot(ArepoSnapshot):
     _fileprefix = "snapshot_"  # underscore is important!
 
     def __init__(self, path, chunksize="auto", catalog=None, **kwargs) -> None:
+        """
+        Initialize an MTNGArepoSnapshot object.
+
+        Parameters
+        ----------
+        path: str
+            Path to the snapshot folder, should contain "output" folder.
+        chunksize: int
+            Chunksize for the data.
+        catalog: str
+            Explicitly state catalog path to use.
+        kwargs:
+            Additional keyword arguments.
+        """
         tkwargs = dict(
             fileprefix=self._fileprefix,
             fileprefix_catalog=self._fileprefix_catalog,
@@ -59,6 +77,22 @@ class MTNGArepoSnapshot(ArepoSnapshot):
     def validate_path(
         cls, path: Union[str, os.PathLike], *args, **kwargs
     ) -> CandidateStatus:
+        """
+        Validate a path as a candidate for the MTNG-Arepo snapshot class.
+
+        Parameters
+        ----------
+        path:
+            Path to validate.
+        args:  list
+        kwargs: dict
+
+        Returns
+        -------
+        CandidateStatus
+            Whether the path is a candidate for this dataset class.
+
+        """
         tkwargs = dict(
             fileprefix=cls._fileprefix, fileprefix_catalog=cls._fileprefix_catalog
         )
@@ -97,6 +131,14 @@ class MTNGArepoCatalog(ArepoCatalog):
     _fileprefix = "fof_subhalo_tab"
 
     def __init__(self, *args, **kwargs):
+        """
+        Initialize an MTNGArepoCatalog object.
+
+        Parameters
+        ----------
+        args: list
+        kwargs: dict
+        """
         kwargs["iscatalog"] = True
         if "fileprefix" not in kwargs:
             kwargs["fileprefix"] = "fof_subhalo_tab"
@@ -107,6 +149,21 @@ class MTNGArepoCatalog(ArepoCatalog):
     def validate_path(
         cls, path: Union[str, os.PathLike], *args, **kwargs
     ) -> CandidateStatus:
+        """
+        Validate a path as a candidate for the MTNG-Arepo catalog class.
+
+        Parameters
+        ----------
+        path: str
+            Path to validate.
+        args: list
+        kwargs: dict
+
+        Returns
+        -------
+        CandidateStatus
+            Whether the path is a candidate for this dataset class.
+        """
         tkwargs = dict(fileprefix=cls._fileprefix)
         tkwargs.update(**kwargs)
         valid = super().validate_path(path, *args, **tkwargs)
