+++ src/scida/customs/gizmo/series.py
@@ -1,3 +1,6 @@
+"""
+Defines a series representing a Gizmo simulation.
+"""
 import os
 
 from scida.customs.gadgetstyle.series import GadgetStyleSimulation
@@ -30,6 +33,21 @@ class GizmoSimulation(GadgetStyleSimulation):
 
     @classmethod
     def validate_path(cls, path, *args, **kwargs) -> CandidateStatus:
+        """
+        Validate a path as a candidate for this simulation class.
+
+        Parameters
+        ----------
+        path: str
+        args: list
+        kwargs: dict
+
+        Returns
+        -------
+        CandidateStatus
+            Whether the path is a candidate for this simulation class.
+
+        """
         if not os.path.isdir(path):
             return CandidateStatus.NO
         if "gizmo_parameters.txt" in os.listdir(path):
