### 2024-01-30 1ab7d5a
add docstrings (#121)
add basic docstrings for most functions and classes
