+++ src/scida/io/_base.py
@@ -1,3 +1,7 @@
+"""
+Basic IO functionality for scida.
+"""
+
 import abc
 import logging
 import os
@@ -22,22 +26,54 @@ log = logging.getLogger(__name__)
 
 
 class InvalidCacheError(Exception):
+    """
+    Raised if a cache file is invalid.
+    """
+
     pass
 
 
 class Loader(abc.ABC):
+    """
+    Base class for loaders.
+    """
+
     def __init__(self, path):
+        """
+        Initialize loader by passing target path.
+        Parameters
+        ----------
+        path: str
+            path to the dataset
+        """
         self.path = path
         self.file = None  # open file handle if available
         self.tempfile = None  # reference to tempfile where needed
 
     @abc.abstractmethod
     def load(self):
+        """
+        Load data from file.
+        Returns
+        -------
+
+        """
         pass
 
 
 class FITSLoader(Loader):
+    """
+    FITS file loader.
+    """
+
     def __init__(self, path):
+        """
+        Initialize loader by passing target path.
+        Parameters
+        ----------
+        path: str
+           path to resource
+        """
         super().__init__(path)
         self.location = self.path
 
@@ -50,6 +86,27 @@ class FITSLoader(Loader):
         virtualcache=False,
         **kwargs
     ):
+        """
+        Load data from FITS file.
+        Parameters
+        ----------
+        overwrite: bool
+            Overwrite cache file if it exists.
+        fileprefix: str
+            Prefix of files to be loaded. If None, we take the first prefix.
+        token: str
+            Token to be used for dask arrays.
+        chunksize: str
+            Chunksize for dask arrays.
+        virtualcache: bool
+            Whether to use virtual caching.
+        kwargs: dict
+            Additional keyword arguments.
+
+        Returns
+        -------
+
+        """
         log.warning("FITS file support is work-in-progress.")
         self.location = self.path
         from astropy.io import fits
@@ -72,7 +129,18 @@ class FITSLoader(Loader):
         return rootcontainer, metadata
 
     def load_metadata(self, **kwargs):
-        """Take a quick glance at the metadata, only attributes."""
+        """
+        Take a quick glance at the metadata, only attributes.
+        Parameters
+        ----------
+        kwargs: dict
+            Additional keyword arguments.
+
+        Returns
+        -------
+        dict
+            Dictionary of attributes.
+        """
         from astropy.io import fits
 
         ext = 0
@@ -80,7 +148,17 @@ class FITSLoader(Loader):
             return hdulist[ext].header
 
     def load_metadata_all(self, **kwargs):
-        """Take a quick glance at metadata."""
+        """
+        Parameters
+        ----------
+        kwargs: dict
+            Additional keyword arguments.
+
+        Returns
+        -------
+        dict
+            Dictionary of attributes.
+        """
         from astropy.io import fits
 
         ext = 1
@@ -90,7 +168,18 @@ class FITSLoader(Loader):
 
 
 class HDF5Loader(Loader):
+    """
+    HDF5 file loader.
+    """
+
     def __init__(self, path):
+        """
+        Initialize loader by passing target path.
+        Parameters
+        ----------
+        path: str
+            path to resource
+        """
         super().__init__(path)
         self.location = ""
 
@@ -103,6 +192,29 @@ class HDF5Loader(Loader):
         virtualcache=False,
         **kwargs
     ):
+        """
+        Load data from HDF5 file.
+        Parameters
+        ----------
+        overwrite: bool
+            Overwrite cache file if it exists.
+        fileprefix: str
+            Prefix of files to be loaded. If None, we take the first prefix.
+        token: str
[…]
