+++ src/scida/fields.py
@@ -1,3 +1,7 @@
+"""
+
+"""
+
 from __future__ import annotations
 
 import inspect
@@ -16,15 +20,41 @@ log = logging.getLogger(__name__)
 
 
 class FieldType(Enum):
-    INTERNAL = 1
-    IO = 2
-    DERIVED = 3
+    """
+    Enum for field types.
+    """
+
+    INTERNAL = 1  # for internal use only
+    IO = 2  # from disk
+    DERIVED = 3  # derived from other fields
 
 
 class FieldRecipe(object):
+    """
+    Recipe for a field.
+    """
+
     def __init__(
         self, name, func=None, arr=None, description="", units=None, ftype=FieldType.IO
     ):
+        """
+        Recipes for a field. Either specify a function or an array.
+
+        Parameters
+        ----------
+        name: str
+            Name of the field.
+        func: Optional[callable]
+            Function to construct array of the field.
+        arr: Optional[da.Array]
+            Array to construct the field.
+        description: str
+            Description of the field.
+        units: Optional[Union[pint.Unit, str]]
+            Units of the field.
+        ftype: FieldType
+            Type of the field.
+        """
         if func is None and arr is None:
             raise ValueError("Need to specify either func or arr.")
         self.type = ftype
@@ -36,7 +66,12 @@ class FieldRecipe(object):
 
 
 class DerivedFieldRecipe(FieldRecipe):
+    """
+    Recipe for a derived field.
+    """
+
     def __init__(self, name, func, description="", units=None):
+        """See FieldRecipe for parameters."""
         super().__init__(
             name,
             func=func,
@@ -63,6 +98,7 @@ class FieldContainer(MutableMapping):
         **kwargs,
     ):
         """
+        Construct a FieldContainer.
 
         Parameters
         ----------
@@ -101,26 +137,71 @@ class FieldContainer(MutableMapping):
         self.internals = ["uid"]  # names of internal fields/groups
         self.parent = parent
 
-    def set_ureg(self, ureg):
+    def set_ureg(self, ureg=None, discover=True):
+        """
+        Set the unit registry.
+
+        Parameters
+        ----------
+        ureg: pint.UnitRegistry
+            Unit registry.
+        discover: bool
+            Attempt to discover unit registry from fields.
+
+        Returns
+        -------
+
+        """
+        if ureg is None and not discover:
+            raise ValueError("Need to specify ureg or set discover=True.")
+        if ureg is None and discover:
+            keys = self.keys(withgroups=False, withrecipes=False, withinternal=True)
+            for k in keys:
+                if hasattr(self[k], "units"):
+                    if isinstance(self[k].units, pint.Unit):
+                        ureg = self[k].units._REGISTRY
         self._ureg = ureg
 
-    def get_ureg(self):
-        keys = self.keys(withgroups=False, withrecipes=False, withinternal=True)
-        for k in keys:
-            if hasattr(self[k], "units"):
-                if isinstance(self[k].units, pint.Unit):
-                    self._ureg = self[k].units._REGISTRY
-        if self._ureg is not None:
-            return self._ureg
-        return None
+    def get_ureg(self, discover=True):
+        """
+        Get the unit registry.
+
+        Returns
+        -------
+
+        """
+        if self._ureg is None and discover:
+            self.set_ureg(discover=True)
+        return self._ureg
 
     def copy_skeleton(self) -> FieldContainer:
+        """
+        Copy the skeleton of the container (i.e., only the containers, not the fields).
+
+        Returns
+        -------
+        FieldContainer
+        """
         res = FieldContainer()
         for k, cntr in self._containers.items():
             res[k] = cntr.copy_skeleton()
         return res
 
[…]
