+++ src/scida/interfaces/mixins/spatial.py
@@ -9,15 +9,30 @@ log = logging.getLogger(__name__)
 
 
 class SpatialMixin(Mixin):
+    """
+    Mixin class for spatial datasets. TBD.
+    """
+
     pass
 
 
 class Spatial3DMixin(SpatialMixin):
+    """Mixin class for 3D spatial datasets. TBD."""
+
     pass
 
 
 class SpatialCartesian3DMixin(Spatial3DMixin):
+    """Mixin class for 3D cartesian spatial datasets."""
+
     def __init__(self, *args, **kwargs):
+        """
+        Initialize the object to add SpatialCartesian3DMixin properties.
+        Parameters
+        ----------
+        args
+        kwargs
+        """
         self.hints = {}
         super().__init__(*args, **kwargs)
         # TODO: determine whether periodic?
@@ -55,16 +70,55 @@ class SpatialCartesian3DMixin(Spatial3DMixin):
                 log.debug("Did not find CoordinatesName for species '%s'" % k)
 
     def get_coords(self, parttype="PartType0"):
+        """
+        Get the coordinates for a given particle type.
+
+        Parameters
+        ----------
+        parttype: str
+            Particle type.
+
+        Returns
+        -------
+        np.ndarray
+            Coordinates.
+        """
         k = self.hints["CoordinatesName"].get(parttype, "")
         if k in self.data[parttype]:
             return self.data[parttype][k]
         return None
 
     def rectangular_cutout_mask(self, center, width, parttype="PartType0"):
+        """
+        Get a rectangular cutout mask for a given particle type.
+
+        Parameters
+        ----------
+        center: np.ndarray
+        width: np.ndarray
+        parttype: str
+
+        Returns
+        -------
+        da.Array
+        """
         coords = self.get_coords(parttype=parttype)
         return rectangular_cutout_mask(
             center, width, coords, pbc=self.pbc, boxsize=self.boxsize
         )
 
     def rectangular_cutout(self, center, width, parttype="PartType0"):
-        pass
+        """
+        Get a rectangular cutout for a given particle type.
+
+        Parameters
+        ----------
+        center: np.ndarray
+        width: np.ndarray
+        parttype: str
+
+        Returns
+        -------
+        da.Array
+        """
+        raise NotImplementedError
