+++ src/scida/registries.py
@@ -1,4 +1,8 @@
-# registries for different object types. mostly automatically registered through __init_subclass__
+"""
+This module contains registries for dataset and dataseries subclasses.
+Subclasses are automatically registered through __init_subclass__
+"""
+
 from typing import Dict, Type
 
 dataset_type_registry: Dict[str, Type] = {}

[…]
