+++ docs/series.md
@@ -2,8 +2,12 @@
 
 !!! info
 
-    If you want to run the code below, you will need to have an AREPO simulation available.
-    Specify the path in load() to the base directory of the simulation, which contains the "output" sub directory.
+
+    If you want to run the code below, you need a folder containing multiple scida datasets as subfolders.
+    Specify the path in load() to the base directory of the series.
+    The example below uses an AREPO simulation, the TNG50-4 simulation, as a series of snapshots.
+    This simulation can be downloaded from the [TNG website](https://www.tng-project.org/data/)
+    or directly accessed online in the [TNGLab](https://www.tng-project.org/data/lab/).
 
 In the tutorial section, we have only considered individual data sets.
 Often data sets are given in a series (e.g. multiple snapshots of a simulation, multiple exposures in a survey).
@@ -11,7 +15,7 @@ Loading this as a series provides convenient access to all contained objects.
 
 ``` pycon
 >>> from scida import load
->>> series = load("TNGvariation_simulation") #(1)!
+>>> series = load("TNG50-4") #(1)!
 ```
 
 1. Pass the base path of the simulation.
