+++ docs/faq.md
@@ -18,7 +18,7 @@ Please note that all fields within a container are expected to have the same sha
 ``` py
 from scida import load
 import dask.array as da
-ds = load('TNG50-4_snapshot')
+ds = load("./snapdir_030")
 array = da.zeros_like(ds.data["PartType0"]["Density"])
 ds.data['PartType0']["zerofield"] = array
 ```
@@ -27,7 +27,7 @@ As we operate with dask, make sure to cast your array accordingly. For example, 
 Alternatively, if you have another dataset loaded, you can assign fields from one to another:
 
 ``` py
-ds2 = load('TNG50-4_snapshot')
+ds2 = load("./snapdir_030")
 ds.data['PartType0']["NewDensity"] = ds2.data['PartType0']["Density"]
 ```
 
