### 2024-02-01 210d4b3
fix doc testing after change of simpaths (#146)
