### 2024-02-01 afeb9ec
add UnitRegistry mismatch to FAQ
