+++ docs/customs/lgalaxies.md
@@ -1,10 +1,0 @@
-# LGalaxies
-
-Access via individual datasets are supported, e.g.:
-
-```pycon
->>> from scida import load
->>> load("LGal_Ayromlou2021_snap58.hdf5")
-```
-
-while access to the series at once (i.e. loading all data for all snapshots in a folder) is **not supported**.
