+++ docs/impressions.md
@@ -1,5 +1,7 @@
 # Visual impressions using scida
 
+The following images were created using scida with matplotlib.
+
 ![Projection from the SIMBA simulation at redshift 2.](images/projection_SIMBA_Temperature.jpg){: style="height:300px"}
 ![Projection from the FLAMINGO simulation at redshift 2.](images/projection_FLAMINGO_Density.jpg){: style="height:300px"}
 ![Projection from the TNG100 simulation at redshift 2.](images/projection_TNG100_GFM_Metallicity.jpg){: style="height:300px"}
