### 2024-02-05 #6064 @mbobra
> I've ticked off the last two items on my copy of the checklist, from my side this passes review.

Thank you for all your time and effort @kyleaoman! Your volunteer time keeps JOSS running and the JOSS team appreciates it very much ☀️
