### 2024-02-13 a06bd88
fix overwrite_cache spell + fix catalog cosmology
