+++ impressions/impressions.md
@@ -1,0 +1,20 @@
+# Visual impressions using scida
+
+The following images were created using scida with matplotlib.
+
+## Simulations
+
+Plots from the SIMBA, FLAMINGO, TNG100 and THESAN simulations.
+
+![Projection from the SIMBA simulation at redshift 2.](images/projection_SIMBA_Temperature.jpg){: style="height:300px"}
+![Projection from the FLAMINGO simulation at redshift 2.](images/projection_FLAMINGO_Density.jpg){: style="height:300px"}
+![Projection from the TNG100 simulation at redshift 2.](images/projection_TNG100_GFM_Metallicity.jpg){: style="height:300px"}
+![Projection from the THESAN simulation at redshift 6.](images/projection_THESAN_NeutralHydrogenAbundance.jpg){: style="height:300px"}
+
+## Observational data
+
+Plots from the SDSS DR16, eROSITA-DE DR1 and Gaia DR3.
+
+![SDSS DR16 Aitoff projection](images/sdss_dr16.png)
+![Data from eROSITA-DE DR1. Mean photon energy.](images/eFEDS.png){: style="height:350px"}
+![Gaia DR3 sky projection](images/simple_hist2d_obs.png){: style="height:350px"}
