### 2024-02-15 #6064 @mbobra
> Passed review from my side as well @mbobra

Thank you so much @EGaraldi! I know it is not easy to make time to volunteer -- your efforts help JOSS tremendously and we really appreciate it 🎉
