### 2024-02-15 #6064 @mbobra
Congratulations, @cbyrohl -- we're almost ready to publish the paper! Could you please cut a new release of scida and deposit it in Zenodo (per [these instructions](https://guides.github.com/activities/citable-code/)) and put the link to the Zenodo deposit in this thread? Please make sure that the author list and title on the Zenodo deposit matches the author list and title on the JOSS paper. Then I can move forward with accepting the paper.
